add form and SQL guards to edit and delete pages

+++ movies/delete-movie.php
@@ -17,10 +17,13 @@
 
   if($_SERVER["REQUEST_METHOD"] == "POST"){
     $movie_ID = mysqli_real_escape_string($conn, $_POST['movie_ID']);
-    $price = mysqli_real_escape_string($conn, $_POST['price']);
     //Add query to insert movie in movies table
     $queryIn = "SELECT * FROM movies where movie_ID = '$movie_ID'";
     $resultIn = mysqli_query($conn, $queryIn);
+    if(mysqli_num_rows($resultIn) == 0){
+      $msg = "Must be an existing ID!";
+    }
+    else {
       $query = "
         DELETE FROM movies
         WHERE movie_ID = $movie_ID
@@ -33,7 +36,7 @@
         echo "ERROR: could not delete the movie $query." . mysqli_error($conn);
       }
     }
-
+  }
   mysqli_close($conn);
 ?>
 
@@ -44,7 +47,7 @@
       <legend>Movie:</legend>
       <p> 
         <label for "movie_ID">Movie ID: </label>
-        <input type = "number" min = 0 max = 999999999 class="required" name = "movie_ID" id = "movie_ID" title= "movie_ID should be numberic">
+        <input type = "number" min = 0 max = 999999999 required name = "movie_ID" id = "movie_ID" title= "movie_ID should be numberic">
       </p>
       <p>
         <input type = "submit" value = "submit"/>
